Fix strict typing under numpy 2.2 stubs and drop dead mypy overrides
The Python 3.10/3.11 CI jobs resolve numpy 2.2, whose stubs predate the
default type parameters on ndarray, so bare np.ndarray annotations and
floating[Any] results from lstsq/linspace fail mypy --strict there while
passing against numpy 2.5 locally. Annotate with NDArray[np.float64] and
cast to float64 so both stub generations type-check.

Also remove the pandas/matplotlib/pyarrow mypy overrides: nothing under
src/ imports them.

+++ src/vacancy_gnn/evaluate.py
@@ -459,7 +459,9 @@ def evaluate(
         An :class:`EvaluationReport` bundling every metric in PLAN.md Section 7.
     """
     if sweep_temperatures is None:
-        sweep_temperatures = np.linspace(0.0, 2.0 * reactor_temperature, 50)
+        sweep_temperatures = np.linspace(0.0, 2.0 * reactor_temperature, 50).astype(
+            np.float64
+        )
 
     return EvaluationReport(
         parity=per_arrangement_parity(model, reference, cutoff=cutoff),
